% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_onset.R
\name{normalize_onset}
\alias{normalize_onset}
\title{A Normalizing Onset Function}
\usage{
normalize_onset(data, nalim = 0.001)
}
\arguments{
\item{data}{Expects a data csv in the form of a CDC fluview submission see FluSight package for a minimal submission}

\item{nalim}{Sets the max probability assigned to no onset. Defaults to 0.001.}
}
\description{
This function allows you to normalize the onset of a CDC fluview submission.
}
\examples{
normalize_onset()
}
\keyword{Normalize}
\keyword{Onset}
