% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_submission.R
\name{normSubmission}
\alias{normSubmission}
\title{A Submission Normalizing Function}
\usage{
normSubmission(data, nalim = 0.001, team = "ISU")
}
\arguments{
\item{data}{Expects a data csv in the form of a CDC fluview submission see FluSight package for a minimal submission}

\item{nalim}{Sets the max probability assigned to no onset. Defaults to 0.001.}

\item{team}{Sets team name for csv submission file name. Defaults to ISU.}
}
\description{
This function will return a CDC fluview submission with the onset probabilities normalized
}
\examples{
normSubmission()
}
\keyword{normalize}
