% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcdcdata.R
\name{plotcdcdata}
\alias{plotcdcdata}
\title{A CDC Data Plot Function}
\usage{
plotcdcdata(region, years = NA)
}
\arguments{
\item{region}{Specifies the region to be plotted.}

\item{years}{Specifies years to retrieve data for (i.e. 2014 for CDC flu season 2014-2015). Default value is the current year and all years values should be > 1997}
}
\description{
This function allows you to plot current and past CDC data.
}
\examples{
plotcdcdata()
}
\keyword{CDC}
\keyword{FluView}
