% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgrid.R
\name{plotgrid}
\alias{plotgrid}
\title{A Grid Plot Function}
\usage{
plotgrid(dat, ilimx)
}
\arguments{
\item{dat}{Expects a data csv in the form of a CDC fluview submission see \code{FluSight} package for a minimal submission}

\item{ilimax}{Numeric. Max level of ILI percentage to be plotted}
}
\description{
This function allows you to plot a grid of plots showing the Week Ahead prediction, Onset, Peak Percentage, Peak Week, and Actual current data.
}
\examples{
plotgrid()
}
\keyword{grid}
\keyword{plots}
