% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeakPer.R
\name{plotPeakPer}
\alias{plotPeakPer}
\title{A Peak Percentage Plot Function}
\usage{
plotPeakPer(dat, region)
}
\arguments{
\item{dat}{Expects a data csv in the form of a CDC fluview submission see \code{FluSight} package for a minimal submission}

\item{region}{Specifies the region to be plotted}
}
\description{
This function allows you to plot Peak Percentage Predictions
}
\examples{
plotPeakPer()
}
\keyword{Peak}
\keyword{Percentage}
\keyword{Plot}
\keyword{Prediction}
