% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeakWeek.R
\name{plotPeakWeek}
\alias{plotPeakWeek}
\title{A Peak Week Plot Function}
\usage{
plotPeakWeek(dat, region)
}
\arguments{
\item{dat}{Expects a data csv in the form of a CDC fluview submission see \code{FluSight} package for a minimal submission}

\item{region}{Specifies the region to be plotted}
}
\description{
This function allows you to plot Peak Week Predictions
}
\examples{
plotPeakWeek()
}
\keyword{Peak}
\keyword{Plot}
\keyword{Prediction}
\keyword{Week}
