% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWeekAhead.R
\name{plotWeekAhead}
\alias{plotWeekAhead}
\title{A Week Ahead Plot Function}
\usage{
plotWeekAhead(dat, region, wk = 1, ilimax)
}
\arguments{
\item{dat}{Expects a data csv in the form of a CDC fluview submission see \code{FluSight} package for a minimal submission}

\item{region}{Specifies the region to be plotted}

\item{wk}{Numeric. How many weeks ahead to plot. Defaults to 1.}

\item{ilimax}{Numeric. Max level of ILI percentage to be plotted}
}
\description{
This function allows you to plot Weak Ahead Predictions
}
\examples{
plotWeekAhead()
}
\keyword{Ahead}
\keyword{Plots}
\keyword{Prediction}
\keyword{Week}
